golden-paths/h3-innovation/mlops-pipeline/skeleton/requirements.txt
azure-identity>=1.15.0
azure-ai-ml>=1.12.0
mlflow>=2.10.0
pandas>=2.0.0
scikit-learn>=1.4.0
